requirements.txt
fastapi[standard]
uvicorn
sqlmodel
python-dotenv
svix-ksuid # used to generate unique k-sorted ids
requests
PyJWT
cryptography